feat(path2): added various modules / split out path2
