+++ src/core/geometry/path2/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  create: require('./create'),
+  fromPoints: require('./fromPoints')
+}
