+++ src/core/geometry/path2/isClosed.js
@@ -1,0 +1,7 @@
+ /**
+   * Determine if the path is a closed or not.
+   * @returns {Boolean} true when the path is closed, otherwise false
+   */
+  const isClosed = path => {
+    return path.closed
+  }
