+++ src/core/geometry/path2/transform.js
@@ -1,0 +1,8 @@
+const transform = (matrix, path) => {
+  let newpoints = path.points.map(function (point) {
+    return point.multiply4x4(matrix4x4)
+  })
+  return new Path2D(newpoints, this.closed)
+}
+
+module.exports = transform
